Split Test logic out of RegressionManager and Scheduler
* Move Test into new `cocotb._test` module.
* Make `cocotb.start_soon` register tasks with a task queue in the Test.
* When Test ends only kill Tasks associated with the Test and not *all*
  Tasks.
* Move cocotb.start_soon, cocotb.create_task, and cocotb.start to the
  new module as they depend upon test-specific knowledge.
* Remove test_complete_cb and test termination logic from Scheduler.
