+++ docs/source/newsfragments/4514.bugfix.rst
@@ -1,0 +1,1 @@
+Fixed bug where scheduled writes at the end of a test are cancelled.
